Get defaults for antibody landscape plotting

This function returns all the options that can be set for plotting of antibody landscapes and the default values for them.

ablandscape.par(
  cex = 1,
  cex.basemap = cex,
  cex.titer = cex,
  col.impulse = "grey20",
  lwd.impulse = 0.5,
  asp.z = 1,
  zaxt = "linear",
  col.grid = "grey80",
  col.surface = "grey20",
  col.surface.grid = "grey20",
  shininess.surface = 30,
  opacity.basemap = 0.4,
  opacity.basemap.ags = opacity.basemap,
  opacity.basemap.sr = opacity.basemap,
  opacity.surface = 0.8,
  opacity.surface.grid = 0.6
)

Value

Returns a list of default plotting options for making antibody landscapes.