Plot an ablandscape contour plot

Plot a contour plot from an antibody landscape fit using `ggplot2` and [ggplot2::geom_contour_filled()].

lndscp_contour_ggplot(
  fit,
  xlim = NULL,
  ylim = NULL,
  crop2chull = TRUE,
  padding = 1,
  grid_x = NULL,
  grid_y = NULL,
  grid_spacing = 0.5,
  show.grid = TRUE,
  grid.color = "grey90",
  grid.linewidth = 0.5,
  ...
)

Arguments

fit
The antibody landscape fit object, created using [ablandscape.fit()]

xlim
x limits for the plot

ylim
y limits for the plot

crop2chull
Crop the plot to the convex hull defined by the antigens titrated

padding
Padding to apply to the fit

grid_x
x grid points at which the ablandscapes fit should be calculated (passed to [predict_lndscp_grid()])

grid_y
y grid points at which the ablandscapes fit should be calculated (passed to [predict_lndscp_grid()])

grid_spacing
grid spacing for the ablandscapes fit (passed to [predict_lndscp_grid()])

show.grid
Show grid lines behind contours

grid.color
Grid line colors when shown

grid.linewidth
Grid line width when shown

...
Additional arguments passed to [ggplot2::geom_contour_filled()]

Value

Returns the ggplot object for the contour plot